Added NotNull filters and Record Count
This commit connects the notNull filters in the front end to the
notnull filters in the back end. It also includes a record count
component on the main page that show the total number of records in
the selected table that match the current filters. The table only
displays the top 200.